move test to right place, fix it and only do this for later json schema versions
