+++ smithy-aws-cloudformation/src/test/resources/software/amazon/smithy/aws/cloudformation/schema/fromsmithy/simple-service-aws.cfn.json
@@ -19,8 +19,7 @@
   },
   "properties": {
     "FooDeprecatedMutableProperty": {
-      "type": "string",
-      "deprecated": true
+      "type": "string"
     },
     "FooId": {
       "type": "string"
